refactor: remove link height, replace logo with component, change data in not found
