+++ src/pages/MainPage/MainPage.module.scss
@@ -48,7 +48,6 @@
 .linkAsButton {
   align-self: center;
   width: clamp(180px, 100%, 240px);
-  height: 66px;
   padding: 20px 30px;
   border: none;
