+++ src/pages/NotFound/NotFound.tsx
@@ -10,28 +10,25 @@ interface NotFoundProps {
   additionalMessage?: string;
 }
 
-export const NotFound: React.FC<NotFoundProps> = memo(
-  // eslint-disable-next-line react/prop-types
-  ({ additionalMessage }) => {
-    return (
-      <main className={cls.notFound}>
-        <div className={cls.headerImg} />
-        <div className={cls.lianaTop}>
-          <Liana />
-        </div>
-        <Logo />
-        <p className={cls.content}>404</p>
-        <p className={cls.signature}>Not Found</p>
-        <AppLink to={Routes.MAIN} className={cls.linkToMain}>
-          To home page
-        </AppLink>
-        {additionalMessage && (
-          <p className={cls.additionalMessage}>{additionalMessage}</p>
-        )}
-        <div className={cls.lianaFooter}>
-          <Liana />
-        </div>
-      </main>
-    );
-  },
-);
+export const NotFound = memo(({ additionalMessage }: NotFoundProps) => {
+  return (
+    <main className={cls.notFound}>
+      <div className={cls.headerImg} />
+      <div className={cls.lianaTop}>
+        <Liana />
+      </div>
+      <Logo />
+      <p className={cls.content}>404</p>
+      <p className={cls.signature}>Not Found</p>
+      <AppLink to={Routes.MAIN} className={cls.linkToMain}>
+        To home page
+      </AppLink>
+      {additionalMessage && (
+        <p className={cls.additionalMessage}>{additionalMessage}</p>
+      )}
+      <div className={cls.lianaFooter}>
+        <Liana />
+      </div>
+    </main>
+  );
+});
