+++ src/pages/RegistrationPage/RegistrationPage.module.scss
@@ -1,6 +1,11 @@
 @import '../../app/styles';
 
 .registrationPage {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
   background-color: $black;
   background-image: url("@/shared/assets/images/reg_background.webp");
   background-position: center;
@@ -8,7 +13,7 @@
   background-blend-mode: overlay;
 }
 
-input{
+input {
   background-color: transparent !important;
 }
 
@@ -38,7 +43,8 @@ input{
   align-items: center;
   justify-content: center;
 
-  .subtitle, .link {
+  .subtitle,
+  .link {
     text-align: center;
 
     @include montserrat(15px, 400, 24px)
@@ -50,14 +56,3 @@ input{
     color: $light-green;
   }
 }
-
-.link__image {
-  transition: 0.2s ease;
-
-  &:hover {
-    cursor: pointer;
-    opacity: 0.7;
-  }
-}
-
-
