+++ src/widgets/Header/Header.module.scss
@@ -10,7 +10,7 @@
   margin: 0 auto;
 }
 
-.wrapper{
+.wrapper {
   display: flex;
   align-items: center;
   justify-content: space-between;
@@ -21,7 +21,7 @@
   border-top-right-radius: 10px;
 }
 
-.header__wrapper{
+.header__wrapper {
   position: relative;
 
   z-index: 10;
@@ -36,33 +36,12 @@
   background-color: $black;
 }
 
-.logo {
-  display: flex;
-  align-items: center;
-
-  transition: .3s;
-
-  &:hover{   
-    transform: rotate(10deg);
-
-    transition: .5s;   
-  }
-
-  @media screen and (width <= 600px) {
-    display: none;
-  }
-}
-
-.logo__image {
-  height: 80%;
-}
-
 .nav {
   display: flex;
   align-items: center;
 }
 
-.nav__list{
+.nav__list {
   display: flex;
   gap: clamp(20px, 4vw, 30px);
 }
@@ -76,26 +55,26 @@
   }
 }
 
-.button__wrapper{
+.button__wrapper {
   display: flex;
   gap: clamp(5px, 4vw, 20px);
   align-items: center;
   height: 100%;
 }
 
-.header__button{
+.header__button {
   height: 60%;
   padding: 5px 15px;
 
   font-size: 20px;
 }
 
-.controls{
+.controls {
   display: flex;
   gap: 20px;
   align-items: center;
 
-  @media screen and (width <= 600px) {
+  @media screen and (width <=600px) {
     justify-content: flex-end;
     width: 100%;
   }
@@ -127,7 +106,7 @@
 }
 
 
-@media screen and (width <= 1100px) {
+@media screen and (width <=1100px) {
   .burger__icon {
     display: flex;
   }
@@ -155,7 +134,7 @@
     align-items: center;
   }
 
-  .nav__link{
+  .nav__link {
     font-size: 25px;
 
     &:hover {
@@ -166,12 +145,11 @@
   }
 
   .mobile__hidden {
-      display: none;
+    display: none;
   }
 
   .mobile__visible {
-      display: flex;
+    display: flex;
   }
 
 }
-
